Extract program builder from check builder.
